backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0

pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3
asyncpg==0.29.0

redis==5.1.1

celery==5.4.0

passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

stripe==10.12.0

httpx==0.27.2
tenacity==9.0.0

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
